Refactor: app.js, server.js 분리 운용
