+++ songinchan/justgramNode/.env
@@ -17,3 +17,5 @@ TYPEORM_LOGGING =TRUE
 
 # secret_key
 SECRET_KEY = 'ThisIsSECRET@key'
+
+PORT = 8000
