+++ songinchan/justgramNode/package.json
@@ -1,7 +1,6 @@
 {
   "name": "justgramnode",
   "version": "1.0.0",
-  "description": "",
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
@@ -13,6 +12,7 @@
   "devDependencies": {
     "bcryptjs": "^2.4.3",
     "express": "^4.18.2",
+    "jest": "^29.2.2",
     "jsonwebtoken": "^8.5.1",
     "nodemon": "^2.0.20"
   },
@@ -22,5 +22,6 @@
     "morgan": "^1.10.0",
     "mysql": "^2.18.1",
     "typeorm": "^0.3.10"
-  }
+  },
+  "description": ""
 }
